combines normalizers
