--- job/scriptsGen/normalizeStepsV1.js
@@ -1,29 +1,0 @@
-'use strict';
-
-var self = normalizeStepsV1;
-module.exports = self;
-
-function normalizeStepsV1(steps) {
-  _.each(steps,
-    function (step) {
-      if (!step.TASK)
-        return;
-
-      // if TASK is in old format convert it to new format
-      if (_.isArray(step.TASK)) {
-        var newTask = {
-          script: []
-        };
-        _.each(step.TASK,
-          function (oldTask) {
-            if (oldTask.script)
-              newTask.script.push(oldTask.script);
-          }
-        );
-        step.TASK = newTask;
-      }
-    }
-  );
-
-  return steps;
-}
